@@ -11,8 +11,8 @@
     "crypto-browserify": "^3.12.0",
     "random": "^3.0.6",
     "random-int": "^3.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "uuid": "^8.3.2",
