@@ -1,17 +1,19 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import { BrowserRouter } from 'react-router-dom';
-import { AppContextProvider } from './context/appContext';
+import React from "react";
+import { createRoot } from "react-dom/client";
 
-ReactDOM.render(
+import { BrowserRouter } from "react-router-dom";
+import { AppContextProvider } from "./context/appContext";
+
+import App from "./App";
+
+import "./index.css";
+
+const root = createRoot(document.getElementById("root"));
+
+root.render(
   <BrowserRouter>
     <AppContextProvider>
       <App />
     </AppContextProvider>
   </BrowserRouter>
-  ,
-  document.getElementById('root')
 );
-
